+++ pyproject.toml
@@ -21,6 +21,7 @@ dependencies = [
     "ga4gh.vrs[extras]~=2.1.2",
     "sqlalchemy~=1.4.54",
     "pyyaml",
+    "pysam",
     "python-dotenv",
     "anyio",
 ]
